pyproject.toml
[tool.poetry]
name = "band_dev"
version = "0.1.0"
description = ""
authors = ["Will Meyers <will@willmeyers.net>"]
readme = "README.md"
packages = [{ include = "band_dev" }]

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.1"
psycopg = "^3.1.17"
base32-crockford = "^0.3.0"
pydantic = { extras = ["email"], version = "^2.5.3" }
django-storages = "^1.14.2"
boto3 = "^1.34.29"
markdown = "^3.5.2"
pyyaml = "^6.0.1"
markdown-full-yaml-metadata = "^2.2.1"
factory-boy = "^3.3.0"
faker = "^23.1.0"
django-environ = "^0.11.2"

[tool.poetry.group.dev.dependencies]
black = "^23.12.1"
django-stubs = "^4.2.7"
mypy = "^1.8.0"
pytest = "^7.4.3"
pylint = "^3.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = ["."]
